Modify: Calculation of status

+++ Backend/CalculationResult.cs
@@ -12,7 +12,7 @@ namespace StatSimulation.Backend
     {
         public string MaxHp { get; set; }
         public string MaxSp { get; set; }
-        public int MaxWeight { get; set; }
+        public decimal MaxWeight { get; set; }
         public string HpRegen { get; set; }
         public string SpRegen { get; set; }
         public string HpRecoveryItem { get; set; }
